+++ tasks.py
@@ -30,6 +30,10 @@ celery = Celery('tasks')
 celery.config_from_object(config_celery)
 
 
+class CeleryException(Exception):
+    pass
+
+
 @celery.task
 def code_deploy(item):
     """
@@ -163,7 +167,7 @@ def site_provision(site):
     :param site: A single site.
     :return:
     """
-    logger.debug('Site provision - {0}'.format(site))
+    logger.debug('Site provision| %s', site)
     start_time = time.time()
     # 'db_key' needs to be added here and not in Eve so that the encryption
     # works properly.
@@ -171,17 +175,35 @@ def site_provision(site):
     # Set future site status for settings file creation.
     site['status'] = 'available'
 
-    provision_task = execute(fabfile.site_provision, site=site)
+    try:
+        logger.debug('Site provision | Create database')
+        result_create_database = execute(fabfile.create_database, site=site)
+    except CeleryException as e:
+        logger.error('Site provision failed | Database creation failed')
+        return result_create_database
 
-    logger.debug(provision_task)
-    logger.debug(provision_task.values)
+    try:
+        provision_task = execute(fabfile.site_provision, site=site)
+        if isinstance(provision_task.get('host_string', None), BaseException):
+            raise provision_task.get('host_string')
+    except CeleryException as e:
+        logger.error('Site provision failed | Error Message | %s', e.message)
 
-    install_task = execute(fabfile.site_install, site=site)
+    logger.debug('Site provision | Provision Fabric task | %s', provision_task)
+    logger.debug('Site provision | Provision Fabric task values | %s', provision_task.values)
 
-    logger.debug(install_task)
-    logger.debug(install_task.values)
+    try:
+        install_task = execute(fabfile.site_install, site=site)
+        if isinstance(install_task.get('host_string', None), BaseException):
+            raise install_task.get('host_string')
+    except CeleryException as e:
+        logger.error('Site install failed | Error Message | %s', e.message)
 
-    patch_payload = {'status': 'available', 'db_key': site['db_key'], 'statistics': site['statistics']}
+    logger.debug('Site provision | Install Fabric task | %s', install_task)
+    logger.debug('Site provision | Install Fabric task values | %s', install_task.values)
+
+    patch_payload = {'status': 'available',
+                     'db_key': site['db_key'], 'statistics': site['statistics']}
     patch = utilities.patch_eve('sites', site['_id'], patch_payload)
 
     profile = utilities.get_single_eve('code', site['code']['profile'])
@@ -191,8 +213,9 @@ def site_provision(site):
     core_string = core['meta']['name'] + '-' + core['meta']['version']
 
     provision_time = time.time() - start_time
-    logger.info('Atlas operational statistic | Site Provision - {0} - {1} | {2} '.format(core_string, profile_string, provision_time))
-    logger.debug('Site has been provisioned\n{0}'.format(patch))
+    logger.info('Atlas operational statistic | Site Provision | %s | %s | %s ',
+                core_string, profile_string, provision_time)
+    logger.debug('Site provision | Patch | %s', patch)
 
     slack_title = '{0}/{1}'.format(base_urls[environment], site['path'])
     slack_link = '{0}/{1}'.format(base_urls[environment], site['path'])
@@ -262,7 +285,7 @@ def site_update(site, updates, original):
 
     if updates.get('status'):
         logger.debug('Found status change.')
-        if updates['status'] in ['installing', 'launching', 'take_down', 'restore']:
+        if updates['status'] in ['installing', 'launching', 'locked', 'take_down', 'restore']:
             if updates['status'] == 'installing':
                 logger.debug('Status changed to installing')
                 # Set new status on site record for update to settings files.
@@ -279,6 +302,9 @@ def site_update(site, updates, original):
                     execute(fabfile.diff_f5)
                     execute(fabfile.update_f5)
                 # Let fabric send patch since it is changing update group.
+            elif updates['status'] == 'locked':
+                logger.debug('Status changed to locked')
+                execute(fabfile.update_settings_file, site=site)
             elif updates['status'] == 'take_down':
                 logger.debug('Status changed to take_down')
                 site['status'] = 'down'
@@ -298,7 +324,8 @@ def site_update(site, updates, original):
                 patch = utilities.patch_eve('sites', site['_id'], patch_payload)
                 logger.debug(patch)
 
-    if updates.get('settings'):
+    # Don't update settings files a second time if status is changing to 'locked'.
+    if updates.get('settings') and updates['status'] != 'locked':
         logger.debug('Found settings change.')
         if updates['settings'].get('page_cache_maximum_age') != original['settings'].get('page_cache_maximum_age'):
             logger.debug('Found page_cache_maximum_age change.')
@@ -473,7 +500,7 @@ def cron(type=None, status=None, include_packages=None, exclude_packages=None):
         site_query_string.append('"status":"{0}",'.format(status))
     else:
         logger.debug('Cron - No status found')
-        site_query_string.append('"status":{"$in":["installed","launched"]},')
+        site_query_string.append('"status":{"$in":["installed","launched","locked"]},')
     if include_packages:
         logger.debug('Cron - found include_packages')
         for package_name in include_packages: